reviewed content to main

+++ common/scripts/mesh-11s.sh
@@ -3,27 +3,48 @@
 function help
 {
     echo
-    echo "Usage: sudo ./mesh_11s.sh <mode> <ip> <mask> <AP MAC> <key> <essid> <freq> <txpower> <country>"
+    echo "Usage: sudo ./mesh_11s.sh <mode> <ip> <mask> <AP MAC> <key> <essid> <freq> <txpower> <country> [<interface>] [<phyname>]"
+    echo "Parameters:"
+    echo "	<mode>"
+    echo "	<ip>"
+    echo "	<mask>"
+    echo "	<AP MAC>"
+    echo "	<WEP key>"
+    echo "	<essid>"
+    echo "	<freq>"
+    echo "	<txpower>"
+    echo "	<country>"
+    echo "	<interface>" - optional
+    echo "	<phyname>"   - optional
     echo
     echo "example:"
-    echo "sudo mesh.sh mesh 192.168.1.2 255.255.255.0 00:11:22:33:44:55 1234567890 mymesh2 5220 30 fi"
+    echo "sudo ./mesh.sh mesh 192.168.1.2 255.255.255.0 00:11:22:33:44:55 1234567890 mymesh2 5220 30 fi wlan1 phy1"
     echo "sudo mesh.sh ap"
     exit
 }
 
-echo "Solving wifi device name.."
-wifidev=$(iw dev | awk '$1=="Interface"{print $2}')
-echo "Found: $wifidev"
+# 1      2    3      4        5     6       7      8         9         10
+# <mode> <ip> <mask> <AP MAC> <key> <essid> <freq> <txpower> <country> <interface>
 
-# 1      2    3      4        5     6       7      8         9
-# <mode> <ip> <mask> <AP MAC> <key> <essid> <freq> <txpower> <country>
+echo "Solving wifi device name.."
+if [[ -z "${10}" ]]; then
+  # one pci qca6174 radio is used in a drone, second one is usb --> can be detected as follows:
+  phyname=$(ls /sys/bus/pci/devices/*/ieee80211/)
+  wifidev=$(ls /sys/class/ieee80211/$phyname/device/net/)
+else
+  wifidev=${10}
+  phyname=${11}
+fi
+echo "Found: $wifidev $phyname"
 
 case "$1" in
 
 mesh)
+
+echo "sudo mesh $1 $2 $3 $4 $5 $6 $7 $8 $9 ${10} ${11}"
       if [[ -z "$1" || -z "$2" || -z "$3" || -z "$4" || -z "$5" || -z "$6" ]]
         then
-          echo "check argumets..."
+          echo "check arguments..."
         help
       fi
 
@@ -46,31 +67,38 @@ network={
 EOF
 
       echo "Killing wpa_supplicant..."
-      killall wpa_supplicant 2>/dev/null
+      # FIXME: If there is another Wi-Fi module being used as an AP for a GW,
+      # this kills that process. We need a better way of handling this. For now
+      # we can just not kill wpa_supplicant when we are loading the mesh_com_tb
+      # module.
+      if [[ -z "${10}" ]]; then
+        killall wpa_supplicant 2>/dev/null
+        rm -fr /var/run/wpa_supplicant/"$wifidev"
+      fi
       killall alfred 2>/dev/null
       killall batadv-vis 2>/dev/null
       rm -f /var/run/alfred.sock
 
 
       echo "unmanage wifi interface.."
-      nmcli dev set $wifidev managed no
+      nmcli dev set "$wifidev" managed no
       modprobe batman-adv
 
       echo "$wifidev down.."
-      iw dev $wifidev del
-      iw phy phy0 interface add $wifidev type ibss
+      iw dev "$wifidev" del
+      iw phy "$phyname" interface add "$wifidev" type ibss
 
       echo "$wifidev create adhoc.."
-      ifconfig $wifidev mtu 1560
+      ifconfig "$wifidev" mtu 1560
 
       echo "$wifidev up.."
-      ip link set $wifidev up
-      batctl if add $wifidev
+      ip link set "$wifidev" up
+      batctl if add "$wifidev"
 
       echo "bat0 up.."
       ifconfig bat0 up
       echo "bat0 ip address.."
-      ifconfig bat0 $2 netmask $3
+      ifconfig bat0 "$2" netmask "$3"
       echo
       ifconfig bat0
 
@@ -82,8 +110,18 @@ EOF
       (batadv-vis -i bat0 -s)&
       echo "started batadv-vis"
 
-      wpa_supplicant -i $wifidev -c /var/run/wpa_supplicant-11s.conf -D nl80211 -C /var/run/wpa_supplicant/ -B
+      # FIXME: Like the comment above - we need to figure out how to handle
+      # multiple Wi-Fi interfaces better. For some reason the background setting
+      # ensures wpa_supplicant doesn't start when this script is run as a process.
+      # This is likely due to the interface not being up in time, and will
+      # require some fiddling with the systemd startup order.
+      if [[ -z "${10}" ]]; then
+        wpa_supplicant -i "$wifidev" -c /var/run/wpa_supplicant-adhoc.conf -D nl80211 -C /var/run/wpa_supplicant/ -B
+      else
+        wpa_supplicant -i "$wifidev" -c /var/run/wpa_supplicant-adhoc.conf -D nl80211 -C /var/run/wpa_supplicant/
+      fi
       ;;
+
 ap)
       if [[ -z "$1" ]]
         then
@@ -91,23 +129,97 @@ ap)
         help
       fi
 
-      # remove all saved wifi networks
-      nmcli --pretty --fields UUID,TYPE con show | grep wifi | cut -b -37 | while read line; do nmcli con delete uuid  $line; done
-
-      nmcli dev set $wifidev managed yes
-      nmcli radio wifi on
-
       # find ap parameters from enclave
       if [ -z "$DRONE_DEVICE_ID" ]
       then
-            echo "DRONE_DEVICE_ID not available"
-            nmcli device wifi hotspot con-name debug-hotspot ssid debug_hotspot band a channel 36 password 1234567890
+        echo "DRONE_DEVICE_ID not available"
+
+cat <<EOF >/var/run/wpa_supplicant-ap.conf
+ctrl_interface=DIR=/var/run/wpa_supplicant
+ap_scan=1
+p2p_disabled=1
+network={
+    ssid="debug-hotspot"
+    mode=2
+    frequency=5220
+    key_mgmt=WPA-PSK
+    psk="1234567890"
+    pairwise=CCMP
+    group=CCMP
+    proto=RSN
+}
+EOF
       else
-            nmcli device wifi hotspot con-name debug-hotspot ssid $DRONE_DEVICE_ID band a channel 36 password 1234567890
+        echo "DRONE_DEVICE_ID available"
+cat <<EOF >/var/run/wpa_supplicant-ap.conf
+ctrl_interface=DIR=/var/run/wpa_supplicant
+ap_scan=1
+p2p_disabled=1
+network={
+    ssid="$DRONE_DEVICE_ID"
+    mode=2
+    frequency=5220
+    key_mgmt=WPA-PSK
+    psk="1234567890"
+    pairwise=CCMP
+    group=CCMP
+    proto=RSN
+}
+EOF
       fi
 
+      echo "Killing wpa_supplicant..."
+      killall wpa_supplicant 2>/dev/null
+      rm -fr /var/run/wpa_supplicant/"$wifidev"
+
+      echo "create $wifidev"
+      iw dev "$wifidev" del
+      iw phy "$phyname" interface add "$wifidev" type managed
+
+      echo "$wifidev up.."
+      ip link set "$wifidev" up
+      batctl if add "$wifidev"
+
+      echo "set ip address.."
+	    # set AP ipaddr
+	    ifconfig "$wifidev" 192.168.1.1 netmask 255.255.255.0
+
+	    # set bat0 ipaddr
+      if [ -z "$DRONE_DEVICE_ID" ]
+        then 
+          # DRONE_DEVICE_ID not available set default
+          ifconfig bat0 192.168.1.1 netmask 255.255.255.0
+        else
+          declare -i ip=10#$(echo "$DRONE_DEVICE_ID" | tr -d -c 0-9)
+          ifconfig bat0 192.168.1."$ip" netmask 255.255.255.0
+      fi
+
+      echo "bat0 up.."
+      ifconfig bat0 up
+      echo
+      ifconfig bat0
+
+      route del -net 192.168.1.0 netmask 255.255.255.0 dev bat0
+      route add -net 192.168.1.0 netmask 255.255.255.0 dev bat0 metric 1
+
+      #TODO
+      # dhserver
+
+      wpa_supplicant -B -i "$wifidev" -c /var/run/wpa_supplicant-ap.conf -D nl80211 -C /var/run/wpa_supplicant/
+
+      ;;
+
+
+off)
+      # service off
+      killall wpa_supplicant 2>/dev/null
+      rm -fr /var/run/wpa_supplicant/"$wifidev"
+      killall alfred 2>/dev/null
+      killall batadv-vis 2>/dev/null
+      rm -f /var/run/alfred.sock 2>/dev/null
       ;;
 *)
       help
       ;;
 esac
+
